header file

+++ 0x0A-argc_argv/main.h
@@ -2,7 +2,7 @@
 #define MAIN_H
 
 void _putchar(char c);
-void _putnbr(unsigned int num);
+void _putnbr(unsigned long int num);
 int _isdigit(char *str);
 
 #endif /* MAIN_H */
